Commit
added comments and changed error with line 14 being la $a0 as it needs to be $s0 so the user input is saved into this value.  jal length is utilized to go the length sub program later to be implemented which should go and get the length of the characters entered.  The length of the character is stored into $s1 and the adress of the last char in the string is saved into $s2.