Commit addition
beqz is used to test if the length of the input is equal to 0 and if it is the program is exited. the move instructions will get the length of the input and the start of a string and use it as passed through parameter for an implemented whitespace function. The first non-white space char it put into $s0 as an address. jal will be used to remove all the leading whitespaces using a function leading_removal that has not been implemented yet.