Commit
userinput is not needed so it was removed. beqz is needed to have as a condition to end the program is the length of string is equal to zero. two parameters are needed to be utilized in ending_removal function that will be implemented later one being the pointer to the end of the string and the other the length of the string. then jump to ending_removal function and return.