Additional Commit
adds to more conditions one being is the length of the input is 0 to end the program by going to invalid function and another is the length is greater than 4 to end the program as is any more than 4 chars we want "Not recognized" to be printed which will be implemented later.  length of the string and the start of the string will be  put into a function to count our sum  and initialize our index.  $a2 is given 29 since that is our base (26+(2936343 %4) = 29