Commit
change function is implemented the index for the load char function is initilized in the change function ($t0). $v0 is the final sum which is returned when jal change is called in the lines above. load char is started I think a equal statement is needed to  make sure the char is valid one to be processed. We can implement that using a function later named valid later.