Commit
added addition equal statement for load char as each char needs to be check to make sure it is valid it is checked by $a0. and if it passes goes to valid function. first loop is also implemented. there needs to be multiple loops and and else statements since we have to remove white spaces from beginning and end but also white spaces in the middle . the firstelse corresponds to the firstloop and will convert the char via multiplication then add to then increment the string pointer the the loop counter. in first loop $t4 is chekced against 1 and if it doesn't equal one then it goes to first else. $t4 should be the index for the exponential loop. We then multiply by $t5 and $a2 being the index and  assigned char respectively. then $t4 being our loop counter is decremented by one and the loop is repeated.