Commit
second loop utilized using the initilized index and checking if it equals one if it does not you go to the secondelse function which has yet to be implemented. if it is not equal to one we then multiply the value by $a2 which is the exponent base of 29.  then we decrement the the exponential loop index and continue until $t4 is equal to one.