Commit
Fixed some formatting. implemented the rangeas function which is the same as the rangeAS function however if the char is greater than s it becomes invalid as there is no accepted char value above s in our program since s is our beta. Like before need to add the regulatory minus addi statement to turn a into 10 which we will do by minusing the char by 87. Then we have to initilize the index for the index exponential loop again and value for the exponent calculation.