Commit
Added second part to the rangeas function like in the rangeAS function that initilizes the index for the exponent loop being our third loop. Where $t1 needs to be minused by 87 since a is 97 when converted to a decimal value but we want it to be 10 in our program. Minusing by 87 turns the "a" into a 10 and all other lowercase values to their appropriate decimal value.  $t5 is load and initiliazed for the exponent calculation to be used in our third loop and thirdelse functions that will be implemented later. $t4 being the result of the subtraction of $a0 and $t0 being the length and load char index which was made in the change function.