Commit
Implemented the thirdloop functions which is the same as the first and second loop corresponding to 0-9 and A-S respectively. $t4 the index for the loop is checked to be 1 and if it is then it goes to the third else to be converted and calculated into the sum. If not then it is multiplied by $a2 being 29 as our base. then the the index is decremented and the loop repeats until $t4 is 1 and heads into the third else statement.