Thirdelse Commit
third else implemented for a-s same as the second and first else statements for 0-9 and A-S respectively. we convert the char then use $v0 to add the conversion to the overall sum. Then increment the index point and the loop index as once we hit 4 that have been added we stop.