Update project2.s
valid function is implemented which gives controls to the function as it copies the data in $ra then jumps to the adress in $ra which is needed as well as the invalid function which has been partly implemented. when invalid is called it needs to throw our the error message "Not Recognized" which is then loaded into $a0.  the invalid function needs to also syscall and to exit afterwards .