Length/noneloop Commit
Implements the length function which finds the length of the input accounting for the increments. $t0 is set to 0 at first for the loops. none loop function is started which at first a null character is checked for with the beqz instruction and it will go to return function wich will store the length and get rid of excess spaces and get the pointer for the last char. it then increments the string pointer $a0 and will also increment the loop $t0 by 1.