Finished function
finished the noneloop function by incrementing the loop and also adding the last jump to noneloop instruction (j noneloop) which continues the loop. return function is implemented to store the length value of $t0to $m0 through use of the move instruction. then it is minused by 1 to get rid of the excess count.