finished function and leading space removal functoin
for the return function gets the pointer for the last char and moves it to $v1. then subtract to from $v1 to get rid of the extra counter. then jr $ra to jump to the adress $ra. Then implemented part of leading_removal function to get the pointer for the 1st non-spaced char and the length of the string without the spaces. it starts out by initializing the index for the loop.