Commit
function for the pointer loop. it starts by loading the start of the string $t0. if $t0 equals $t2 (decimal value for space =32) it increments and if its equal to $t3 (decimal value for tab =0) it also increments. it also uses the beq instruction to see if it is equal to $a0 in which case it goes to the invalid function and ends the program. while printing out the error message.