Fixed
added nonwhitespacereturn function for the pointerloop as it was added for the trailing spaces but not the procedding spaces. Same as the lastcharreturn however it stores the index point of the first non space char and then stores the length of the input minus the leading increments. also added a jump to nonwhitespacereturn in the pointer loop to perform the function.